requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
pillow==10.4.0
django-cors-headers==4.4.0
django-phonenumber-field==8.0.0
